+++ CHANGES.rst
@@ -271,6 +271,8 @@ API Changes
     Astroquery 0.3.5 and will be removed from Astropy in a future version.
     [#5558, #5904]
 
+  - The `astropy.vo.samp` package no longer supports SSL.
+
 - ``astropy.wcs``
 
   - Removed deprecated ``wcs.rotateCD``. [#6170]